Merge 1.3 changes upstream

+++ Generator/Hints/HintTypes/JunkHint.cs
@@ -54,7 +54,7 @@ namespace TPRandomizer.Hints
         public static JunkHint decode(
             HintEncodingBitLengths bitLengths,
             BitsProcessor processor,
-            Dictionary<int, byte> itemPlacements
+            Dictionary<int, int> itemPlacements
         )
         {
             ushort idValue = (ushort)processor.NextInt(16);
